+++ lib/librte_ether/rte_ethdev.c
@@ -3628,7 +3628,7 @@ rte_eth_dev_adjust_nb_rx_tx_desc(uint16_t port_id,
 }
 
 int
-rte_eth_dev_pool_ops_supported(uint8_t port_id, const char *pool)
+rte_eth_dev_pool_ops_supported(uint16_t port_id, const char *pool)
 {
 	struct rte_eth_dev *dev;
